+++ playlist_bridge/__init__.py
@@ -1,4 +1,4 @@
 """Playlist Bridge 2.0 package."""
 
 __version__ = "2.0.0-beta.1"
-__build__ = "20260906.4"
+__build__ = "20260908.6"